explicitely provide the port name in ExportedLink
The scheme of using either link name or port name was actually a bad
idea, as it is not stable across changes in the SDF file and/or changes
in the transformer's frame names. It would therefore force changing
the network configuration each time e.g. a typo is fixed ;-)

Moreover, this splits configuration resolution from the creation of the
actual ports in setupLinks to avoid leaking ports when the configuration
is bad.