workaround the lack of mercator projection in Gazebo's spherical coordinate handling
The GPSTask component now optionally allows to use Gazebo's own
SphericalCoordinates class to handle the latlon-to-cartesian conversion,
ensuring that the position matches the vehicle "true" position.

Long-run, we should propose a mercator projection to Gazebo. The
current method basically assumes that a vehicle that stays at Z=0
starts flying after a while ...